+++ build-logic/src/main/kotlin/MetroProjectExtension.kt
@@ -101,7 +101,6 @@ constructor(private val project: Project, objects: ObjectFactory) {
           iosSimulatorArm64()
           iosArm64()
           iosX64()
-          macosX64()
         } else {
           // Tier 1
           macosArm64()
@@ -120,9 +119,6 @@ constructor(private val project: Project, objects: ObjectFactory) {
           // Tier 3
           mingwX64()
           iosX64()
-          macosX64()
-          tvosX64()
-          watchosX64()
           if (!requiresAndroidXDeps) {
             androidNativeArm32()
             androidNativeArm64()
@@ -160,20 +156,12 @@ constructor(private val project: Project, objects: ObjectFactory) {
             it.platformType == KotlinPlatformType.js || it.platformType == KotlinPlatformType.wasm
           }
           .configureEach {
-            val target = this
             compilations.configureEach {
               compileTaskProvider.configure {
                 compilerOptions {
                   freeCompilerArgs.add(
                     "-Xklib-duplicated-unique-name-strategy=allow-all-with-warning"
                   )
-                  //                  if (target.platformType == KotlinPlatformType.js) {
-                  //                    freeCompilerArgs.add(
-                  //                      // These are all read at compile-time
-                  //
-                  // "-Xwarning-level=RUNTIME_ANNOTATION_NOT_SUPPORTED:disabled"
-                  //                    )
-                  //                  }
                 }
               }
             }